Fix: update to the same type comparison

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -10,7 +10,7 @@ int main(void)
 	char value[] = "_putchar\n";
 	int i;
 
-	for (i = 0; i < strlen(value); i++)
+	for (i = 0; i < (int) strlen(value); i++)
 	{
 		_putchar(*(value + i));
 	}
